Add APP SDK 3.0 bugfix for clImage tests
Found that with the ROCm software stack, the AMD OpenCL runtime
sometimes failed when testing 3D images. These create a huge
number of events, and we would randomly segfault on the GPU
while waiting on these events. A workaround is to enqueue a
marker on those events and wait on its output event.