Benchmark installation and run fixes
Fixed a number of benchmark installation and run problems:
1. For versions of AMD OpenCL that do not run on the AMD APP
SDK using the Catalyst driver stack, we removed the need to
compile and run benchmarks that use C++ in the kernels;
these are not supported and would cause things like the
auto-tester to fail.

2. Removed LINPACK on versions of AMD's OpenCL that do not
use the Catalyst AMD APP SDK, as it will not work. In
addition, do not build and attempt to run LINPACK if not using
an AMD "Hawaii" GPU, as this is the only GPU supported by our
current LINPACK installation. This entails removing LINPACK from
the full "ALL_BENCHMARKS" list and only running it if requested
specifically.

3. Patch the CG benchmark in the SNU OpenCL version of the NAS
Parallel Benchmarks. It has a number of intra-kernel data races
caused by multiple workgroups trying to update the same global
memory locations during the parallel matrix generation phase.
This leads to non-deterministic output. Worse still, we observed
that this could cause buffer underflows when some matrix offset
values were set to -1 and then used to index into arrays. Our
fix is to fall back to the serial CPU version that generates
the matrices on the host.