Bypass bad OpenCL runtime handling of OpenCL images
Some OpenCL runtimes (in particular, the ROCm OpenCL runtime with
"Hawaii" GPUs) incorrectly handle some image functions in such a
way that clARMOR will lock up when running them.

Rather than trying to debug the runtime, we simply skip checks of
images when we detect these situations. Updated our runtime to
find this case and update the testbench to work around it.