Fix for Hawaii+ROCm GPU check lockups.
When using a Hawaii GPU on the ROCm software stack, we end up
causing lockups somewhere in the command queue if we put our
checks to go onto the GPU.

We can still successfully check for buffer overflows on cl_mem
and SVM buffers if we force the checks onto the host. So we do that.