Update install scripts to support 'super'
On the ROCm software stack, we want to install the 'super' utility
so that we can run ROCm-smi with it. This allows people in the
'video' group, who can use the GPUs, to run rocm-smi without needing
to be in the sudoers group and without needing to type in their
sudoers password every time.

Installing 'super' on all the other Ubuntu 16.04.3 systems as well
to maintain parity.