Fix Retain/Release problem
We did not create the kinfo structure for a kernel until its first
argument was set. However, cl.hpp (for example) will call clRetain()
and clRelease() repeatedly on a kernel while setting its arguments.

For example, it calls clRetain() before setting the kernel arg. We
would ignore that clRetain(), because the kinfo hadn't been created
yet. Then the argument would be set, and we would create kinfo.
The following clRelease would then delete the kinfo. This meant that
after setting a bunch of kernel arguments, we would have no kinfo
and would potentially segfault when dereferencing it.

First: this patch fixes the segfault to a more clean crash.
Second: this patch creates a kinfo structure when we call clRetain()
and sets its reference count to the proper value so that clARMOR
can work with kernels using cl.hpp