Image API improvement when images not supported
On OpenCL devices where images are not supported, we previously
did not cleanly handle calls that should have returned errors
because of the lack of images. Instead, we would die inside the
wrapper. These changes fix that in the OpenCL wrapper, so that such
APIs check for image support and reeturn proper OpenCL errors
to the user.

This also updates the tests to properly check for image support
before they call any Image-expecting APIs.