Fix version number printing, Python 3 printing
The ctypes Python function that allows us to pull the version
number out of the clARMOR library could fail to work because the
pointer is returned, by default, as an integer type. Instead, we
need to set the restype of the library function so that it is
returned as a c_char_p. This is important for cross-compat with
the 64-bit library.

This change should also fix up the clARMOR script prints to work
with Python 3. In addition, we should now properly prepend the
'clARMOR: ' string on all outputs